Update README — fix terminology (#218)
I think the term was used completely wrong here. A "render callback" is becoming commonly accepted as a term for describing when a component is passed a function as children (or via an explicit `render` prop).

Here I believe the consideration is to just give advice to those willing to avoid creating new functions when the render method is called. I also chose a more neutral phrasing since this is sort of a controversial topic of whether you actually gain much by avoiding these.